Initializing repositories with default methods

+++ src/main/java/org/example/rodeodrivediner_webapp/repositories/CartRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Long> {
+public interface CartRepository extends JpaRepository<Cart, Integer> {
 
     Cart findCartByCustomer(Customer customer);
     boolean existsCartByCustomer(Customer customer);
