+++ src/main/java/org/example/rodeodrivediner_webapp/repositories/ProductInCartRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductInCartRepository extends JpaRepository<ProductInCart, Long> {
+public interface ProductInCartRepository extends JpaRepository<ProductInCart, Integer> {
     ProductInCart findByCart(Cart cart);
     ProductInCart findByProducts(Product prodotto);
     boolean existsByProducts(Product prodotto);
