+++ src/main/java/org/example/rodeodrivediner_webapp/repositories/PurchaseRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
+public interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
 
     List<Purchase> findByCustomer(Customer customer);
